Fix tests for #12376 Refactor and add tests (#12472)
